+++ components/input/input_default.story.vue
@@ -1,6 +1,7 @@
 <template>
   <dt-input
-    :value="value"
+    ref="input"
+    v-model="inputValue"
     :type="type"
     :messages="messages"
     :size="size"
@@ -12,11 +13,16 @@
     :messages-class="messagesClass"
     :placeholder="placeholder"
     :input-class="inputClass"
+    :current-length="currentLength"
+    :validate="validationConfig"
     @blur="onBlur"
     @input="onInput"
     @clear="onClear"
+    @focus="onFocus"
     @focusin="onFocusIn"
     @focusout="onFocusOut"
+    @update:length="updateLength"
+    @update:invalid="onUpdateIsInvalid"
   >
     <template
       v-if="labelSlot"
@@ -51,7 +57,56 @@ import icon from '@/common/mixins/icon';
 
 export default {
   name: 'InputDefault',
+
   components: { DtInput },
+
   mixins: [icon],
+
+  data () {
+    return {
+      inputValue: '',
+      inputLength: 0,
+    };
+  },
+
+  computed: {
+    validationMessage () {
+      const remainingCharacters = this.validate?.length?.max - this.inputLength;
+
+      if (remainingCharacters < 0) {
+        return `${Math.abs(remainingCharacters)} characters over limit`;
+      } else {
+        return `${remainingCharacters} characters left`;
+      }
+    },
+
+    validationConfig () {
+      if (!this?.validate?.length) {
+        return null;
+      }
+
+      // Deep clone validate object
+      const validateConfigData = JSON.parse(JSON.stringify(this.validate));
+
+      // Adds validation message
+      validateConfigData.length.message = this?.validate?.length?.message
+        ? this.validate.length.message : this.validationMessage;
+
+      return validateConfigData;
+    },
+  },
+
+  watch: {
+    value (val) {
+      this.inputValue = val;
+    },
+  },
+
+  methods: {
+    updateLength ($event) {
+      this.inputLength = $event;
+      this.onUpdateLength($event);
+    },
+  },
 };
 </script>